Renamed cookieroutes to cookieinfoRoutes
